new updates

+++ src/stories_title/Stories.service.ts
@@ -2,13 +2,20 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Story } from './entites/Stories.Entity';
 import { Repository } from 'typeorm';
 import { CreateStoryDTO } from './DTOs/CreateStory.DTO';
-import { BadRequestException, Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  ForbiddenException,
+  Injectable,
+} from '@nestjs/common';
 import { EventsGateway } from '../events/events.gateway';
+import { JwtService } from '@nestjs/jwt';
+import { JWT_Payload } from 'src/utils';
 @Injectable()
 export class StoriesSrvices {
   constructor(
     @InjectRepository(Story) private readonly storyrepo: Repository<Story>,
     private eventGetway: EventsGateway,
+    private readonly jwtServices: JwtService,
   ) {}
 
   public async getAllStories() {
@@ -47,10 +54,14 @@ export class StoriesSrvices {
     return this.storyrepo.save(newStory);
   }
 
-  public async deleteStory(id: number) {
+  public async deleteStory(id: number, payload: JWT_Payload) {
     const story = await this.storyrepo.findOne({ where: { id: id } });
 
     if (story) {
+      if (story.creator !== payload.username) {
+        throw new ForbiddenException('You are not allowed to do this');
+      }
+
       const newStory = await this.storyrepo.remove(story);
       this.eventGetway.sendUpdate('storydeleted', newStory);
       return `${newStory.title} has been deleted`;
